+++ tests/blockchain/Test_BlockchainAPI.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2018 The Open-Transactions developers
+// Copyright (c) 2019 The Open-Transactions developers
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
@@ -242,7 +242,10 @@ public:
 
             EXPECT_EQ(expectedCount, thread.item().size());
 
-            if (expectedCount > thread.item().size()) { return false; }
+            if (expectedCount >
+                static_cast<std::size_t>(thread.item().size())) {
+                return false;
+            }
 
             auto i = int{0};
